refactor: remove `NativeStorage`/`WebStorage`
