+++ jestSetup.js
@@ -1,6 +1,6 @@
 jest.mock('./lib/storage');
-jest.mock('./lib/storage/platforms/NativeStorage', () => require('./lib/storage/__mocks__'));
-jest.mock('./lib/storage/platforms/WebStorage', () => require('./lib/storage/__mocks__'));
+jest.mock('./lib/storage/platforms/index.native', () => require('./lib/storage/__mocks__'));
+jest.mock('./lib/storage/platforms/index', () => require('./lib/storage/__mocks__'));
 jest.mock('./lib/storage/providers/IDBKeyValProvider', () => require('./lib/storage/__mocks__'));
 
 jest.mock('react-native-device-info', () => ({getFreeDiskStorage: () => {}}));
