--- lib/storage/platforms/NativeStorage.ts
@@ -1,3 +1,0 @@
-import {default as NativeStorage} from '../providers/SQLiteProvider';
-
-export default NativeStorage;
