--- lib/storage/platforms/WebStorage.ts
@@ -1,3 +1,0 @@
-import {default as WebStorage} from '../providers/IDBKeyValProvider';
-
-export default WebStorage;
