+++ lib/storage/platforms/index.native.ts
@@ -1,3 +1,3 @@
-import NativeStorage from './NativeStorage';
+import {default as NativeStorage} from '../providers/SQLiteProvider';
 
 export default NativeStorage;
