+++ lib/storage/platforms/index.ts
@@ -1,3 +1,3 @@
-import WebStorage from './WebStorage';
+import {default as WebStorage} from '../providers/IDBKeyValProvider';
 
 export default WebStorage;
