+++ tests/unit/storage/providers/StorageProviderTest.js
@@ -1,11 +1,11 @@
 /* eslint-disable import/first */
-jest.unmock('../../../../lib/storage/platforms/NativeStorage');
-jest.unmock('../../../../lib/storage/platforms/WebStorage');
+jest.unmock('../../../../lib/storage/platforms/index.native');
+jest.unmock('../../../../lib/storage/platforms/index');
 jest.unmock('../../../../lib/storage/providers/IDBKeyValProvider');
 
 import _ from 'underscore';
-import NativeStorage from '../../../../lib/storage/platforms/NativeStorage';
-import WebStorage from '../../../../lib/storage/platforms/WebStorage';
+import NativeStorage from '../../../../lib/storage/platforms/index.native';
+import WebStorage from '../../../../lib/storage/platforms/index';
 
 it('storage providers have same methods implemented', () => {
     const nativeMethods = _.keys(NativeStorage);
